chore: Make linter happy.

+++ evb/src/model.h
@@ -4,6 +4,6 @@
 #include "arm_math.h"
 
 void init_model(void);
-int model_inference(float32_t *x, uint32_t *y);
+int model_inference(float32_t *x, float32_t *y);
 
 #endif // __MODEL_H
